pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "torch>=2.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ltx-video"
version = "0.1.2"
description = "A package for LTX-Video model"
authors = [
    { name = "LTX-Video Team", email = "ltx-video@lightricks.com" }
]
requires-python = ">=3.10"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dynamic = ["dependencies"]

[project.optional-dependencies]
inference = [
    "imageio[ffmpeg]",
    "av",
    "torchvision"
]
test = [
    "pytest",
]

[tool.setuptools.packages.find]
include = ["ltx_video*"]

[tool.setuptools.package-data]
ltx_video = ["configs/*.yaml", "**/*.cpp", "**/*.cu"]
